finished error handling on player & team, evaluationTable determine rank for empty array, better display of set settings, tournaments shown on player & team detail with working links ==> added getTournamentType() to all tournament classes

+++ src/main/java/cz/tournament/control/service/TeamService.java
@@ -2,7 +2,9 @@ package cz.tournament.control.service;
 
 import cz.tournament.control.domain.Participant;
 import cz.tournament.control.domain.Team;
+import cz.tournament.control.domain.Tournament;
 import cz.tournament.control.domain.User;
+import cz.tournament.control.domain.exceptions.ParticipantInTournamentException;
 import cz.tournament.control.repository.ParticipantRepository;
 import cz.tournament.control.repository.TeamRepository;
 import cz.tournament.control.repository.UserRepository;
@@ -26,12 +28,15 @@ public class TeamService {
     private final UserRepository userRepository;
     private final TeamRepository teamRepository;
     private final ParticipantRepository participantRepository;
+    private final TournamentService tournamentService;
 
-    public TeamService(UserRepository userRepository, TeamRepository teamRepository, ParticipantRepository participantRepository) {
+    public TeamService(UserRepository userRepository, TeamRepository teamRepository, ParticipantRepository participantRepository, TournamentService tournamentService) {
         this.userRepository = userRepository;
         this.teamRepository = teamRepository;
         this.participantRepository = participantRepository;
+        this.tournamentService = tournamentService;
     }
+    
     
     public Team createTeam(Team team){
         //set creator as user
@@ -54,23 +59,36 @@ public class TeamService {
         return result;
     }
     
+    @Transactional(readOnly = true)
     public List<Team> findAll(){
         List<Team> teams = teamRepository.findByUserIsCurrentUser();
         return teams;
     }
     
+    @Transactional(readOnly = true)
     public Team findOne(Long id){
         Team team = teamRepository.findOneWithEagerRelationships(id);
         return team;
+    }
+    
+    @Transactional(readOnly = true)
+    public List<Tournament> findAllTournaments(Team team){
+        Participant participant = participantRepository.findByTeam(team);
+        return tournamentService.findByParticipant(participant);
     }
     
     /**
      * Deletes a team entity and its associated participant entity.
      * 
      * @param id of entity to delete
+     * @throws cz.tournament.control.domain.exceptions.ParticipantInTournamentException
      */
-    public void delete(Long id){
+    public void delete(Long id) throws ParticipantInTournamentException{
         Participant participant = participantRepository.findByTeam(teamRepository.findOne(id));
+        List<Tournament> hisTournaments = tournamentService.findByParticipant(participant);
+        if(!hisTournaments.isEmpty()){
+            throw new ParticipantInTournamentException("Cannot delete participant: "+participant.toString()+", is in at least one tournament.");
+        }
         teamRepository.delete(id);
         participantRepository.delete(participant);
     }
